BackendEvent: change x,y to float
Expressed as points, not pixels (on macOS at least, I think on
other platforms too?)

+++ src/opencreatures1/PointerManager.h
@@ -7,8 +7,8 @@ class PointerManager {
   public:
 	ObjectHandle m_pointer_tool;
 
-	int m_screenx;
-	int m_screeny;
+	float m_screenx;
+	float m_screeny;
 
 	PointerManager() {}
 	void update();
